feat(wallet-dashboard): redirect to /dashboard and remove homepage
