+++ apps/wallet-dashboard/next.config.mjs
@@ -6,6 +6,11 @@ const nextConfig = {
     async redirects() {
         return [
             {
+                source: '/',
+                destination: '/dashboard/home',
+                permanent: true,
+            },
+            {
                 source: '/dashboard',
                 destination: '/dashboard/home',
                 permanent: true,
